paramterize old copying manager tests
