+++ tests/unit/copying_manager_tests/common.py
@@ -938,6 +938,8 @@ _TestableCopyingManagerWorkerProxy = multiprocessing.managers.MakeProxyType(  # 
         six.ensure_str("get_checkpoints_path"),
         six.ensure_str("_init_scalyr_client"),
         six.ensure_str("change_agent_log"),
+        six.ensure_str("get_checkpoints"),
+        six.ensure_str("write_checkpoints"),
     ],
 )
