+++ tests/unit/copying_manager_tests/test_environment.py
@@ -21,14 +21,14 @@ import shutil
 import tempfile
 
 if False:
-    from typing import Union
     from typing import Dict
-    from typing import Tuple
     from typing import Optional
+    from typing import List
 
 from scalyr_agent.configuration import Configuration
-from scalyr_agent.log_processing import LogMatcher, LogFileProcessor
 from scalyr_agent.platform_controller import DefaultPaths
+
+from scalyr_agent import scalyr_logging
 
 
 import six
@@ -47,21 +47,10 @@ class TestableLogFile(object):
 
     __test__ = False
 
-    def __init__(self, config, name):
-        # type: (Configuration, Union[str, pathlib.Path]) -> None
+    def __init__(self, path):
+        # type: (pathlib.Path) -> None
 
-        self._config = config
-        self.name = name
-
-        # the path of the file, it is unknown until the file is created by the 'create' function
-        self.path = None  # type: Optional[pathlib.Path]
-
-    def initialize(self, root_path):
-        """
-        Determine the path of the file and Create file in the root path.
-        """
-        self.path = pathlib.Path(root_path, self.name)
-        self.create()
+        self._path = pathlib.Path(path)
 
     def create(self):
         self.path.touch()
@@ -78,83 +67,17 @@ class TestableLogFile(object):
                 file.write(line)
                 file.write("\n")
 
-    def _create_log_matcher_for_log_file(self):  # type: () -> LogMatcher
-        """
-        Create log matcher from the appropriate log config from the configuration.
-        """
-
-        log_config = next(
-            iter(lc for lc in self._config.log_configs if lc["path" == self.path])
-        )
-
-        return LogMatcher(self._config, log_config)
-
-    def create_single_log_processor(
-        self, checkpoints=None, copy_at_index_zero=False
-    ):  # type: (Dict, bool) -> LogFileProcessor
-        """
-        Create log file processor by the log matcher.
-        :return:
-        """
-
-        matcher = self._create_log_matcher_for_log_file()
-
-        if checkpoints is None:
-            checkpoints = {}
-
-        processors = []
-        for processor in matcher.find_matches(
-            existing_processors=[],
-            previous_state=checkpoints,
-            copy_at_index_zero=copy_at_index_zero,
-        ):
-            processors.append(processor)
-
-        return processors[0]
+    @property
+    def path(self):
+        return self._path
 
     @property
     def str_path(self):
         return six.text_type(self.path)
 
 
-def after_initialize(f):
-    """
-    Decorator for the methods of the 'ConfigBuilder' class.
-    It raises error if decorated method must not be called before Config builder is initialized.
-    """
-
-    def wrapper(self, *args, **kwargs):
-        if not self._initialized:
-            raise RuntimeError(
-                "Method '{0}' cannot be called before this Test config is initialized.".format(
-                    f.__name__
-                )
-            )
-
-        return f(self, *args, **kwargs)
-
-    return wrapper
-
-
-def before_initialize(f):
-    """
-    Decorator for the methods of the 'ConfigBuilder' class.
-    It raises error if decorated method must not be called after Config builder is initialized.
-    """
-
-    def wrapper(self, *args, **kwargs):
-        if self._initialized:
-            raise RuntimeError(
-                "Method '{0}' cannot be called after this Test config is initialized.".format(
-                    f.__name__
-                )
-            )
-        return f(self, *args, **kwargs)
-
-    return wrapper
-
-
 class TestingConfiguration(Configuration):
+    __test__ = False
     """
     A subclass of the original configuration class, that allow to assign additional options
     which will be useful for testing.
@@ -176,53 +99,31 @@ class TestEnvironBuilder(object):
     It allows to make additions which will be included to the resulting config file.
     """
 
-    def __init__(self, config_data=None):
+    MAX_NON_GLOB_TEST_LOGS = 10
+
+    def __init__(self, config_data=None, root_path=None):
 
         self._config_initial_data = config_data
 
         self._config = None  # type: Optional[TestingConfiguration]
-        self._cleared = None
 
-        self._root_path = None
+        self._root_path = root_path
 
-        # self._log_processors = dict()  # type: Dict[str, LogFileProcessor]
-
-        self._initialized = False
-
-        # additional log files that should be reflected in the "logs" section of the configuration.
-        self._log_files = {}  # type: Dict[str, TestableLogFile]
-
-        self.__use_pipelining = None
+        self._current_test_log_files = []
 
     def __del__(self):
         self.clear()
 
-    def initialize(self):
-        """
-        Creates config object and prepares needed environment.
-        This config will contain all additions that are made before this method is called.
-        :return:
-        """
+    def init_agent_dirs(self):
+        # if self._root_path is None:
+        self._root_path = pathlib.Path(tempfile.mkdtemp(prefix="scalyr_testing"))
+        self.agent_data_path.mkdir()
+        self.agent_logs_path.mkdir()
+        self.test_logs_dir.mkdir()
+        self.glob_logs_dir.mkdir()
+        self.non_glob_logs_dir.mkdir()
 
-        self._root_path = root_path = pathlib.Path(
-            tempfile.mkdtemp(prefix="scalyr_testing")
-        )
-        self._data_dir_path = root_path / "data"
-        self._data_dir_path.mkdir()
-
-        self._logs_dir_path = root_path / "logs"
-        self._logs_dir_path.mkdir()
-
-        self._agent_config_path = root_path / "agent.json"
-
-        for log_file in self._log_files.values():
-            log_file.initialize(self._root_path)
-
-        self._create_config(self._config_initial_data)
-
-        self._initialized = True
-
-    def _create_config(self, config_data=None):  # type: (Optional[Dict]) -> None
+    def init_config(self, config_data=None):  # type: (Optional[Dict]) -> None
         """
         Create config object and apply all additions.
         :param config_data:
@@ -230,9 +131,9 @@ class TestEnvironBuilder(object):
         """
 
         default_paths = DefaultPaths(
-            six.text_type(self._logs_dir_path),
-            six.text_type(self._agent_config_path),
-            six.text_type(self._data_dir_path),
+            six.text_type(self.agent_logs_path),
+            six.text_type(self.agent_config_path),
+            six.text_type(self.agent_data_path),
         )
 
         if config_data is None:
@@ -246,138 +147,112 @@ class TestEnvironBuilder(object):
 
         logs = config_data.get("logs", list())
 
-        for log_file in self._log_files.values():
-            logs.append({"path": six.text_type(log_file.path)})
+        logs.append({"path": six.text_type(self.glob_logs_dir / "log_*.log")})
+
+        for i in range(self.MAX_NON_GLOB_TEST_LOGS):
+            path = self.non_glob_logs_dir / ("log_%s.log" % i)
+            logs.append({"path": six.text_type(path)})
 
         config_data["logs"] = logs
 
-        self._agent_config_path.write_text(six.text_type(json.dumps(config_data)))
+        self.agent_config_path.write_text(six.text_type(json.dumps(config_data)))
 
         config = TestingConfiguration(
-            six.text_type(self._agent_config_path), default_paths, None
+            six.text_type(self.agent_config_path), default_paths, None
         )
 
         config.parse()
 
         self._config = config
 
-    @property  # type: ignore
-    @after_initialize
-    def config(self):  # type: () -> Optional[Configuration]
-        return self._config
+    def recreate_files(self, files_number, path, file_name_format=None):
+        # type: (int, pathlib.Path, six.text_type) -> List
+        """
+        Recreate *file_number* of files in the *path* directory with the *file_name_format* name format.
+        :return:
+        """
 
-    @property  # type: ignore
-    @after_initialize
-    def log_files(self):  # type: () -> Dict[str, TestableLogFile]
-        return self._log_files
+        if file_name_format is None:
+            file_name_format = "log_{0}.log"
 
-    @after_initialize
+        self.remove_files(path)
+        result = []
+        for i in range(files_number):
+            test_file = TestableLogFile(path / file_name_format.format(i))
+            test_file.create()
+            self._current_test_log_files.append(test_file)
+            result.append(test_file)
+
+        return result
+
+    def remove_files(self, path):
+        # type: (pathlib.Path) -> None
+        for log_path in path.iterdir():
+            log_path.unlink()
+
+        self._current_test_log_files = []
+
+    @property
+    def config(self):  # type: () -> Configuration
+        return self._config  # type: ignore
+
     def get_log_config(self, log_file):  # type: (TestableLogFile) -> Dict
         """
         Get log config from existing config.
         """
 
-        for log_config in self._config.log_configs:  # type: ignore
-            if six.text_type(log_file.path) == log_config["path"]:
-                return log_config
+        log_config = self.config.parse_log_config({"path": log_file.str_path})
 
-        raise RuntimeError(
-            "Log config for file: '{0}' is not found.".format(log_file.path)
-        )
+        return dict(log_config)
 
     def clear(self):
-        if not self._cleared:
+        scalyr_logging.close_handlers()
+
+        if self._root_path.exists():
             shutil.rmtree(six.text_type(self._root_path))
-            self._cleared = True
 
-    @before_initialize
-    def add_log_file(
-        self, name=None, create=True
-    ):  # type: (Optional[str], bool) -> TestableLogFile
-        """
-        Add new file into "logs" section.
-        :param name: Name to the file relative to ths environment root path.
-        :param create: Create file in filesystem if True
-        """
+    @property
+    def root_path(self):
+        return self._root_path
 
-        if name is None:
-            name = "test_file_{0}".format(len(self._log_files))
-
-        file_obj = TestableLogFile(config=self._config, name=name)  # type: ignore
-
-        self._log_files[name] = file_obj
-
-        return file_obj
-
-    @after_initialize
-    def get_log_file(self, name):  # type: (str) -> TestableLogFile
-        log_file = self._log_files.get(name)
-        if log_file is None:
-            raise RuntimeError(
-                "Can not find file '{0}' in config".format(log_file.path)  # type: ignore
-            )
-
-        return log_file  # type: ignore
-
-    @after_initialize
-    def get_log_file_path(self, name):  # type: (str) -> pathlib.Path
-
-        log_file = self._log_files.get(name)
-        if log_file is None:
-            raise RuntimeError("Can not find file '{0}' in config")
-
-        return pathlib.Path(log_file.path)  # type: ignore
-
-    @after_initialize
-    def append_lines_to_log_file(self, name, *lines):
-        path = self.get_log_file_path(name)
-        with path.open("a") as file:
-            for line in lines:
-                file.write(line)
-                file.write("\n")
-
-    @classmethod
-    def create_with_n_files(cls, n, config_data=None):
-        # type: (int, Dict) -> Tuple[Tuple[TestableLogFile, ...], TestEnvironBuilder]
-        """
-        Convenient config factory which creates config builder with n log_files.
-        """
-        test_environment = cls(config_data=config_data)
-
-        log_files = tuple(test_environment.add_log_file(create=True) for _ in range(n))
-
-        test_environment.initialize()
-        return log_files, test_environment
-
-    @property  # type: ignore
-    @after_initialize
+    @property
     def agent_data_path(self):
-        return self._data_dir_path
+        return self.root_path / "data"
 
-    @property  # type: ignore
-    @after_initialize
-    def agent_log_path(self):
-        return self._logs_dir_path
+    @property
+    def agent_logs_path(self):
+        return self.root_path / "logs"
 
-    @after_initialize
+    @property
+    def agent_config_path(self):
+        return self.root_path / "agent.json"
+
+    @property
+    def test_logs_dir(self):
+        return self._root_path / "test-logs"
+
+    @property
+    def glob_logs_dir(self):
+        return self.test_logs_dir / "glob-logs"
+
+    @property
+    def non_glob_logs_dir(self):
+        return self.test_logs_dir / "non-glob-logs"
+
     def get_checkpoints_path(self, worker_id):  # type: (six.text_type) -> pathlib.Path
-        return self._data_dir_path / ("checkpoints-%s.json" % worker_id)
+        return self.agent_data_path / ("checkpoints-%s.json" % worker_id)
 
-    @after_initialize
     def get_active_checkpoints_path(
         self, worker_id
     ):  # type: (six.text_type) -> pathlib.Path
-        return self._data_dir_path / "active-checkpoints-%s.json" % worker_id
+        return self.agent_data_path / "active-checkpoints-%s.json" % worker_id
 
-    @after_initialize
     def get_checkpoints(self, worker_id):
         return json.loads(self.get_checkpoints_path(worker_id).read_text())
 
-    @after_initialize
     def get_active_checkpoints(self, worker_id):
         return json.loads(self.get_active_checkpoints_path(worker_id).read_text())
 
-    @property  # type: ignore
-    @after_initialize
-    def root_path(self):
-        return self._root_path
+    @property
+    def current_test_log_files(self):
+        return self._current_test_log_files
